Update BHMF in components.py
The mbh_mass_func did not handle scatter appropriately and is now depricated. This function is replaced with mbh_mass_func_conv, which performs a convolution between the GSMF and MMBulge relation and is far more accurate. It still underestimates number density for the highest masses, but this is negligible for scatter_dex > 0.2